Move Target to command.rs
This feels slightly more appropriate.
